Rename things + more documentation
