+++ biome.json
@@ -7,6 +7,7 @@
       "package.json",
       "biome.json",
       "*.config.ts",
+      "vitest.*.ts",
       "tsconfig*.json",
       "testing/*.mjs"
     ],
