+++ src/components/Pacs/mergeStates.ts
@@ -21,9 +21,9 @@ type SeriesQueryZip = {
  * Fragments of the state of a DICOM series exists remotely in three places:
  * PFDCM, CUBE, and LONK.
  *
- * Join the states from those places into one mega-object.
+ * Merge the states from those places into one mega-object.
  */
-function joinStates(
+function mergeStates(
   pfdcm: ReadonlyArray<StudyAndSeries>,
   cubeSeriesQuery: ReadonlyArray<SeriesQueryZip>,
   receiveState: ReceiveState,
@@ -59,35 +59,23 @@ function joinStates(
 }
 
 /**
- * State coalescence.
- *
- * It is assumed that ChRIS has the following behavior for each DICOM series:
- *
- * 1. ChRIS_ui subscribes to a series' notifications via LONK
- * 2. ChRIS_ui checks CUBE whether a series exists in CUBE
- * 3. When both subscription and existence check is complete,
- *    and the series does not exist in CUBE, ChRIS_ui is ready
- *    to pull the DICOM series.
- * 4. During the reception of a DICOM series, `status.done === false`
- * 5. After the reception of a DICOM series, ChRIS enters a "waiting"
- *    state while the task to register the DICOM series is enqueued
- *    or running.
- * 6. The DICOM series will appear in CUBE after being registered.
+ * State coalescence for the "PACS Retrieve Workflow" described in the
+ * tsdoc for {@link PacsController}.
  */
 function pullStateOf(
   state: SeriesReceiveState,
-  result?: { isLoading: boolean; data: any },
+  cubeQueryResult?: { isLoading: boolean; data: any },
 ): SeriesPullState {
-  if (!result) {
+  if (!cubeQueryResult) {
     // request to check CUBE whether series exists has not been initiated
     return SeriesPullState.NOT_CHECKED;
   }
-  if (!state.subscribed || result.isLoading) {
+  if (!state.subscribed || cubeQueryResult.isLoading) {
     // either not subscribed yet, or request to check CUBE whether series
     // exists is pending
     return SeriesPullState.CHECKING;
   }
-  if (result.data === null) {
+  if (cubeQueryResult.data === null) {
     // checked, series DOES NOT exist in CUBE
     if (state.done) {
       // finished receiving by oxidicom, waiting for CUBE to register
@@ -102,4 +90,4 @@ function pullStateOf(
 
 export type { SeriesQueryZip };
 export { pullStateOf };
-export default joinStates;
+export default mergeStates;
